refactor(health): split complexity walker into focused modules
walker.py had grown to ~1.8k lines spanning eight distinct concerns. Split
it into one module per pass, leaving walker.py as a thin orchestrator that
parses the source once and drives the passes:

- models.py         output dataclasses
- ast_utils.py      name/text helpers, function-node collection, params
- nloc.py           non-blank / non-comment line counting
- cyclomatic.py     the CCN / cognitive / nesting engine
- assertions.py     assertion-block detection
- error_handling.py error-handling anti-pattern detection
- perf_walk.py      the performance-risk pass
- class_analysis.py class-level LCOM4 / god-class metrics

Pure code-motion, no behavior change. The package __init__ still re-exports
the dataclasses and walk_file / walk_file_complexity, and walker.py keeps the
re-exports tests reach for, so every importer is unaffected. The shared
identifier-suffix constant moves to ast_utils so assertions and class_analysis
both import it instead of relying on a cross-section forward reference.

Full unit suite green (5730 passed).
